+++ src/core/frameworks/index.ts
@@ -1,0 +1,2 @@
+export { applySvelteKitRules } from "./sveltekitRules.js";
+export { applyNextJsRules } from "./nextJsRules.js";